requirements.txt
Django==1.11
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
